Fix the put when you want to add alergies and method and get single patient

+++ apps/patients/views.py
@@ -56,15 +56,21 @@ class PatientList(APIView):
 
     def put(self, request, pk):
         patient = self.get_object(pk)
-        serializer = PatientSerializer(patient, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
+        if patient is None:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+        else:
+            print(patient.first())
+            serializer = PatientSerializer(patient.first(), data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
 
     def delete(self, request, pk):
         patient = self.get_object(pk)
+        if patient is None:
+            return Response(status=status.HTTP_404_NOT_FOUND)
         patient.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
